pyproject.toml
[project]
name = "medical-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.10",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.1",
    "pypdf>=6.7.4",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
]
